feat: tweaked the way we do Sundays so that there is only one rule for them, and they must match on the year too.
