+++ ReadingsBuilder/ReadingsBuilder/Model/Pipeline/Steps/Step10ManuallyExtractedSundays.cs
@@ -1,0 +1,26 @@
+﻿
+using ReadingsBuilder.Model.Data;
+using ReadingsBuilder.Model.Pipeline.DTOs;
+using ReadingsBuilder.Model.Pipeline.Steps.Utility;
+
+namespace ReadingsBuilder.Model.Pipeline.Steps
+{
+    public class Step10ManuallyExtractedSundays : BaseStep, IStep
+    {
+        private readonly IByDayOfMonthRuleSetApplier ruleSetApplier;
+
+        public Step10ManuallyExtractedSundays(IRuleApplier ruleApplier, IAllDataFactory dataFactory, IByDayOfMonthRuleSetApplier ruleSetApplier) : base(ruleApplier, dataFactory)
+        {
+            this.ruleSetApplier = ruleSetApplier;
+        }
+
+        public int Order => 10;
+
+        protected override string RuleSetName => "ManuallyExtractedSundays.cs";
+
+        public PipelineWorkingResult RunStep(PipelineWorkingResult workingResult)
+        {
+            return ruleSetApplier.ApplyRulesByDayOfMonth(workingResult, ApplicableRules);
+        }
+    }
+}
